# api/openapi.yaml
openapi: 3.0.3
info:
  title: ManyRouter Management API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - operatorToken: []
paths:
  /healthz:
    get:
      operationId: getHealth
      security: []
      responses:
        '200':
          description: Service health
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /sites:
    post:
      operationId: createSite
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
      responses:
        '201':
          description: Site created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        default:
          $ref: '#/components/responses/Error'
  /sites/{site_id}:
    get:
      operationId: getSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Site details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        default:
          $ref: '#/components/responses/Error'
  /suppliers:
    post:
      operationId: createSupplier
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '201':
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        default:
          $ref: '#/components/responses/Error'
  /suppliers/{supplier_id}:
    get:
      operationId: getSupplier
      parameters:
        - $ref: '#/components/parameters/SupplierId'
      responses:
        '200':
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        default:
          $ref: '#/components/responses/Error'
  /site-suppliers:
    post:
      operationId: createSiteSupplier
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteSupplierRequest'
      responses:
        '201':
          description: Site supplier relation and initial route plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteSupplierResponse'
        default:
          $ref: '#/components/responses/Error'
  /site-suppliers/{relation_id}:
    get:
      operationId: getSiteSupplier
      parameters:
        - $ref: '#/components/parameters/RelationId'
      responses:
        '200':
          description: Site supplier state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteSupplierResponse'
        default:
          $ref: '#/components/responses/Error'
  /site-suppliers/{relation_id}/sync:
    post:
      operationId: syncSiteSupplier
      parameters:
        - $ref: '#/components/parameters/RelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Synchronization accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncOperationResponse'
        default:
          $ref: '#/components/responses/Error'
  /sync-operations/{operation_id}:
    get:
      operationId: getSyncOperation
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          description: Synchronization state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncOperationResponse'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    SiteId:
      name: site_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SupplierId:
      name: supplier_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RelationId:
      name: relation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OperationId:
      name: operation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
    CreateSiteRequest:
      type: object
      required: [code, name, new_api_base_url, new_api_access_token]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 63
        name:
          type: string
          minLength: 1
          maxLength: 120
        new_api_base_url:
          type: string
          format: uri
          maxLength: 2048
        new_api_access_token:
          type: string
          minLength: 1
          maxLength: 8192
    SiteResponse:
      type: object
      required: [id, code, name, new_api_base_url, status, compatibility_status, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        new_api_base_url:
          type: string
          format: uri
        status:
          type: string
          enum: [enabled, disabled]
        compatibility_status:
          type: string
          enum: [unknown, compatible, incompatible]
        version:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SupplierModelRequest:
      type: object
      required: [model, upstream_model, input_price, output_price, currency]
      properties:
        model:
          type: string
          minLength: 1
          maxLength: 191
        upstream_model:
          type: string
          minLength: 1
          maxLength: 191
        input_price:
          type: string
          pattern: '^\\d+(\\.\\d{1,10})?$'
        output_price:
          type: string
          pattern: '^\\d+(\\.\\d{1,10})?$'
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
    SupplierModelResponse:
      allOf:
        - $ref: '#/components/schemas/SupplierModelRequest'
        - type: object
          required: [enabled]
          properties:
            enabled:
              type: boolean
    CreateSupplierRequest:
      type: object
      required: [code, name, upstream_base_url, upstream_api_key, models]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 63
        name:
          type: string
          minLength: 1
          maxLength: 120
        upstream_base_url:
          type: string
          format: uri
          maxLength: 2048
        upstream_api_key:
          type: string
          minLength: 1
          maxLength: 8192
        models:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/SupplierModelRequest'
    SupplierResponse:
      type: object
      required: [id, code, name, protocol, upstream_base_url, status, version, models, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        protocol:
          type: string
          enum: [openai_compatible]
        upstream_base_url:
          type: string
          format: uri
        status:
          type: string
          enum: [draft, enabled, disabled]
        version:
          type: integer
          format: int64
        models:
          type: array
          items:
            $ref: '#/components/schemas/SupplierModelResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateSiteSupplierRequest:
      type: object
      required: [site_id, supplier_id, group_display_name, sale_ratio]
      properties:
        site_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        group_display_name:
          type: string
          minLength: 1
          maxLength: 120
        sale_ratio:
          type: string
          pattern: '^(0|[1-9]\\d*)(\\.\\d{1,6})?$'
        visible:
          type: boolean
          default: true
          description: Whether the dedicated entry is open. Closing it also rejects later requests from existing API keys bound to this group after synchronization.
    SiteSupplierResponse:
      type: object
      required: [id, site_id, supplier_id, group_key, group_display_name, sale_ratio, visible, desired_status, sync_status, route_plan_version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        group_key:
          type: string
        group_display_name:
          type: string
        sale_ratio:
          type: string
        visible:
          type: boolean
          description: Whether the dedicated entry is open, including access by existing group-bound API keys.
        desired_status:
          type: string
          enum: [observing, enabled, disabled]
        sync_status:
          type: string
          enum: [pending, syncing, active, failed, manual_locked]
        route_plan_version:
          type: integer
          format: int64
        last_confirmed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SyncOperationResponse:
      type: object
      required: [id, relation_id, route_plan_id, status, attempt, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        relation_id:
          type: string
          format: uuid
        route_plan_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, uncertain, succeeded, retryable_failed, manual_required]
        current_step:
          type: string
          nullable: true
        attempt:
          type: integer
        last_error_code:
          type: string
          nullable: true
        last_error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true